requirements.txt
asgiref
black
certifi
cffi
charset-normalizer
click
cloudinary
coreapi
coreschema
cryptography
Django
django-autoslug
django-cors-headers
django-countries
django-filter
django-js-asset
django-mptt
django-phonenumber-field
djangorestframework
djangorestframework-simplejwt
drf-yasg
idna
inflection
itypes
Jinja2
Markdown
MarkupSafe
mypy-extensions
packaging
pathspec
phonenumbers
platformdirs
pycparser
PyJWT
pytz
requests
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
tomli
typing_extensions
uritemplate
urllib3
